# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Songguo Admin API
  version: "1"
  description: >
    Control-plane (admin) API for the Songguo AI usage gateway. It is read-mostly
    and consumed by the dashboard, the `songguo` MCP server, and any other tooling.
    Every endpoint is gated by a single admin bearer key (`SONGGUO_ADMIN_KEY`);
    when that key is empty the server runs the API unprotected and logs a warning.

    This describes only the control plane. The data plane — the transparent proxy
    at the native vendor paths (`/v1/*`, `/api/v3/*`), with the provider selected
    by the `X-Songguo-Provider` header, else the body's model, else the wire
    default — speaks each vendor's native wire protocol and is documented in
    docs/registry.md.
servers:
  - url: /
security:
  - adminKey: []
...
paths:
  /api/overview:
    get:
      tags: [usage]
      summary: Spend & health summary over a window
      operationId: getOverview
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Overview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/series:
    get:
      tags: [usage]
      summary: Cost/request/error totals bucketed over time
      operationId: getUsageSeries
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - name: bucket
          in: query
          schema: { type: string, enum: [hour, day] }
          description: Bucket size; omit to auto-select (day for ranges over 2 days, else hour).
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/tokens-by-model:
    get:
      tags: [usage]
      summary: Per-bucket tokens broken down by model, plus total cost
      operationId: getUsageTokensByModel
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - name: bucket
          in: query
          schema: { type: string, enum: [hour, day] }
          description: Bucket size; omit to auto-select (day for ranges over 2 days, else hour).
      responses:
        "200":
          description: Tokens-by-model series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokensByModelSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/success-by-model:
    get:
      tags: [usage]
      summary: Per-bucket request/error counts broken down by a dimension
      operationId: getUsageSuccessByModel
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - name: bucket
          in: query
          schema: { type: string, enum: [hour, day] }
          description: Bucket size; omit to auto-select (day for ranges over 2 days, else hour).
        - name: dimension
          in: query
          schema: { type: string, enum: [model, vendor, user] }
          description: Breakdown dimension; defaults to model.
      responses:
        "200":
          description: Success-by-model series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessByModelSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/cache-by-model:
    get:
      tags: [usage]
      summary: Per-bucket cache-read and total-input tokens broken down by a dimension
      operationId: getUsageCacheByModel
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - name: bucket
          in: query
          schema: { type: string, enum: [hour, day] }
          description: Bucket size; omit to auto-select (day for ranges over 2 days, else hour).
        - name: dimension
          in: query
          schema: { type: string, enum: [model, vendor, user] }
          description: Breakdown dimension; defaults to model.
      responses:
        "200":
          description: Cache-by-model series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CacheByModelSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/context/composition:
    get:
      tags: [usage]
      summary: Aggregated context-window decomposition over a window
      operationId: getContextComposition
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Context composition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextComposition" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/breakdown:
    get:
      tags: [usage]
      summary: Per-dimension usage breakdown (model, vendor, user, or modality)
      operationId: getUsageBreakdown
      parameters:
        - name: dimension
          in: query
          required: true
          schema: { type: string, enum: [model, vendor, user, modality] }
          description: Column to group by.
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Breakdown
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Breakdown" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/errors:
    get:
      tags: [usage]
      summary: Error-row counts grouped by class over a window
      operationId: getUsageErrors
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Error counts by class
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageErrors" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/usage/error-codes:
    get:
      tags: [usage]
      summary: Top error rows grouped by upstream status code over a window
      operationId: getUsageErrorCodes
      parameters:
        - name: dimension
          in: query
          schema: { type: string, enum: [model, vendor, user], default: model }
          description: Series dimension the optional key filter applies to.
        - { name: key, in: query, schema: { type: string }, description: "Scope to one series (e.g. one model); empty means all traffic." }
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Error counts by upstream status code, ranked by count (top 8)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageErrorCodes" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/feed:
    get:
      tags: [calls]
      summary: Activity feed — one row per session (aggregated) or standalone request
      operationId: listFeed
      parameters:
        - { name: user_id, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: vendor, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer }, description: HTTP status filter (per-call, before grouping) }
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: A page of feed rows plus the total group count
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/sessions/overview:
    get:
      tags: [calls]
      summary: Aggregate coding-agent session stats over a window
      operationId: getSessionsOverview
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Session aggregate stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/calls:
    get:
      tags: [calls]
      summary: Browse the per-call ledger (filtered, paginated)
      operationId: listCalls
      parameters:
        - { name: user_id, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: vendor, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer }, description: HTTP status filter }
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: A page of calls plus the total count for the filter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CallsPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/calls/export:
    get:
      tags: [calls]
      summary: Download filtered calls as CSV or JSON
      operationId: exportCalls
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, json], default: csv }
        - { name: user_id, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: vendor, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer } }
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: A downloadable attachment (text/csv or application/json)
          content:
            text/csv: { schema: { type: string } }
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Call" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/calls/{id}:
    get:
      tags: [calls]
      summary: A single call entry by id
      operationId: getCall
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: The call entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Call" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/calls/{id}/trace:
    get:
      tags: [calls]
      summary: Captured request/response payload for a call
      operationId: getCallTrace
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Trace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/sessions/{id}:
    get:
      tags: [calls]
      summary: One session's rollups, agent tree, and calls
      operationId: getSession
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/sessions/{id}/messages:
    get:
      tags: [calls]
      summary: Compact prompt material reconstructed from captured session requests
      operationId: getSessionMessages
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: De-duplicated system, tool, and message values in their native wire shape
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionMessages" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/sessions/{id}/context:
    get:
      tags: [calls]
      summary: One session's per-turn context-window composition
      operationId: getSessionContext
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Session context composition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionContext" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users:
    get:
      tags: [users]
      summary: List users with computed lifetime spend
      operationId: listUsers
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [users]
      summary: Create a user (plaintext key returned once)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUser" }
      responses:
        "201":
          description: The created user, including the plaintext `key`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}:
    get:
      tags: [users]
      summary: Get one user with computed lifetime spend
      operationId: getUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [users]
      summary: Update a user's mutable fields
      operationId: patchUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchUser" }
      responses:
        "200":
          description: The updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete a user (the admin user cannot be deleted)
      operationId: deleteUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: User deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/{id}/revoke:
    post:
      tags: [users]
      summary: Revoke a user's key
      operationId: revokeUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The revoked user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vendors:
    get:
      tags: [providers]
      summary: List vendors from the live snapshot (with stats)
      operationId: listVendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Vendor" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/vendors/{name}/test:
    post:
      tags: [providers]
      summary: Probe a vendor host for reachability
      operationId: testVendor
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Reachability result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/me:
    get:
      tags: [meta]
      summary: Whoami — the caller's role and, for a consumer key, its scope and budget
      operationId: getMe
      responses:
        "200":
          description: Identity
          content:
            application/json:
              schema:
                type: object
                required: [role, id, name, scope]
                properties:
                  role: { type: string, enum: [admin, user] }
                  id: { type: string }
                  name: { type: string }
                  scope: { type: array, items: { type: string } }
                  budget: { type: number, nullable: true }
                  spend: { type: number }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/services:
    get:
      tags: [services]
      summary: Auto-derived, model-centric services
      operationId: listServices
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Service" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/providers:
    get:
      tags: [providers]
      summary: List configured providers (keys masked)
      operationId: listProviders
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Provider" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [providers]
      summary: Create a provider
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProvider" }
      responses:
        "201":
          description: The created provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/providers/{id}:
    get:
      tags: [providers]
      summary: Get one provider
      operationId: getProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [providers]
      summary: Update a provider's mutable fields
      operationId: patchProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchProvider" }
      responses:
        "200":
          description: The updated provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [providers]
      summary: Delete a provider
      operationId: deleteProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/providers/{id}/test:
    post:
      tags: [providers]
      summary: Probe a provider host for reachability
      operationId: testProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Reachability result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/catalog:
    get:
      tags: [config]
      summary: Provider preset catalogue
      operationId: getCatalog
      responses:
        "200":
          description: Catalogue (preset directory)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    adminKey:
      type: http
      scheme: bearer
      description: The admin key from SONGGUO_ADMIN_KEY.
  parameters:
    Since:
      name: since
      in: query
      schema: { type: integer, format: int64 }
      description: Window start, unix seconds.
    Until:
      name: until
      in: query
      schema: { type: integer, format: int64 }
      description: Window end, unix seconds.
  responses:
    Unauthorized:
      description: Missing or invalid admin key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    UsageSeries:
      type: object
      properties:
        bucket: { type: string }
        points: { type: array, items: { $ref: "#/components/schemas/SeriesPoint" } }
...